[rename] 폴더명 수정 #179
